refactor: reorganize indicators package structure and update imports and documentation
