+++ tests/unit/indicators/test_snow.py
@@ -12,7 +12,7 @@ import pytest
 import xarray
 from pytest_mock import MockerFixture
 
-from earthkit.climate.indicators.land import snow
+from earthkit.climate.indicators import snow
 
 INDICATORS = [
     (snow.blowing_snow, "blowing_snow", {"val": "test"}),
